# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/cv-analysis":{"post":{"summary":"Cv Analysis","operationId":"cv_analysis_cv_analysis_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CVAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CVAnalysis"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/grammar-analysis":{"post":{"summary":"Grammar Analysis","operationId":"grammar_analysis_grammar_analysis_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrammarAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrammarAnalysis"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/meal-plan":{"post":{"summary":"Meal Plan","operationId":"meal_plan_meal_plan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MealPlanRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MealPlan"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CVAnalysis":{"properties":{"work_experience":{"type":"string","title":"Work Experience"},"skills":{"type":"string","title":"Skills"},"professional_summary":{"type":"string","title":"Professional Summary"}},"type":"object","required":["work_experience","skills","professional_summary"],"title":"CVAnalysis"},"CVAnalysisRequest":{"properties":{"job_title":{"type":"string","title":"Job Title"},"experience_years":{"type":"string","title":"Experience Years","default":""},"skills":{"type":"string","title":"Skills","default":""},"current_role":{"type":"string","title":"Current Role","default":""},"job_description":{"type":"string","title":"Job Description"},"cv_text":{"type":"string","title":"Cv Text"}},"type":"object","required":["job_title","job_description","cv_text"],"title":"CVAnalysisRequest"},"GrammarAnalysis":{"properties":{"grammatical_errors":{"type":"string","title":"Grammatical Errors","default":"None"},"syntactic_errors":{"type":"string","title":"Syntactic Errors","default":"None"},"general_suggestions":{"type":"string","title":"General Suggestions"},"rewrite":{"type":"string","title":"Rewrite"}},"type":"object","required":["general_suggestions","rewrite"],"title":"GrammarAnalysis"},"GrammarAnalysisRequest":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"GrammarAnalysisRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MealPlan":{"properties":{"day1":{"type":"string","title":"Day1"},"day2":{"type":"string","title":"Day2"},"day3":{"type":"string","title":"Day3"},"day4":{"type":"string","title":"Day4"},"day5":{"type":"string","title":"Day5"},"day6":{"type":"string","title":"Day6"},"day7":{"type":"string","title":"Day7"}},"type":"object","required":["day1","day2","day3","day4","day5","day6","day7"],"title":"MealPlan"},"MealPlanRequest":{"properties":{"meal_title":{"type":"string","title":"Meal Title"},"user_preferences":{"$ref":"#/components/schemas/UserPreferences"},"nutritional_goals":{"$ref":"#/components/schemas/NutritionalGoals"}},"type":"object","required":["meal_title","user_preferences","nutritional_goals"],"title":"MealPlanRequest"},"NutritionalGoals":{"properties":{"caloric_intake":{"type":"integer","title":"Caloric Intake"}},"type":"object","required":["caloric_intake"],"title":"NutritionalGoals"},"UserPreferences":{"properties":{"dietary_restrictions":{"type":"string","title":"Dietary Restrictions"},"allergies":{"type":"string","title":"Allergies"},"favorite_foods":{"type":"string","title":"Favorite Foods"},"disliked_foods":{"type":"string","title":"Disliked Foods"},"mealPlan_purpose":{"type":"string","title":"Mealplan Purpose"},"country":{"type":"string","title":"Country"}},"type":"object","required":["dietary_restrictions","allergies","favorite_foods","disliked_foods","mealPlan_purpose","country"],"title":"UserPreferences"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}